refactor default graph converter
