+++ unravel/utils/objects/default_graph_converter.py
@@ -1,5 +1,6 @@
 import logging
 import sys
+import inspect
 
 from dataclasses import dataclass, field, asdict
 
@@ -22,6 +23,10 @@ from ..features import (
 
 from .default_graph_settings import DefaultGraphSettings
 from .custom_spektral_dataset import CustomSpektralDataset
+
+from ..features.utils import make_sparse, reshape_from_size
+
+import numpy as np
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
@@ -164,17 +169,63 @@ class DefaultGraphConverter:
     def _convert(self):
         raise NotImplementedError()
 
-    def to_graph_frames(self) -> dict:
-        raise NotImplementedError()
+    def to_spektral_graphs(self) -> List[Graph]:
+        if not self.graph_frames:
+            self.to_graph_frames()
 
-    def to_pickle(self) -> None:
-        raise NotImplementedError()
+        return [
+            Graph(
+                x=d["x"],
+                a=d["a"],
+                e=d["e"],
+                y=d["y"],
+                id=d["id"],
+            )
+            for d in self.graph_frames
+        ]
+
+    def to_pickle(self, file_path: str, verbose: bool = False) -> None:
+        """
+        We store the 'dict' version of the Graphs to pickle each graph is now a dict with keys x, a, e, and y
+        To use for training with Spektral feed the loaded pickle data to CustomDataset(data=pickled_data)
+        """
+        if not file_path.endswith("pickle.gz"):
+            raise ValueError(
+                "Only compressed pickle files of type 'some_file_name.pickle.gz' are supported..."
+            )
+
+        if not self.graph_frames:
+            self.to_graph_frames()
+
+        if verbose:
+            print(f"Storing {len(self.graph_frames)} Graphs in {file_path}...")
+
+        import pickle
+        import gzip
+        from pathlib import Path
+
+        path = Path(file_path)
+
+        directories = path.parent
+        directories.mkdir(parents=True, exist_ok=True)
+
+        with gzip.open(file_path, "wb") as file:
+            pickle.dump(self.graph_frames, file)
 
     def to_spektral_graphs(self) -> List[Graph]:
         if not self.graph_frames:
             self.to_graph_frames()
 
-        return [g.to_spektral_graph() for g in self.graph_frames]
+        return [
+            Graph(
+                x=d["x"],
+                a=d["a"],
+                e=d["e"],
+                y=d["y"],
+                id=d["id"],
+            )
+            for d in self.graph_frames
+        ]
 
     def to_custom_dataset(self) -> CustomSpektralDataset:
         """
@@ -182,3 +233,73 @@ class DefaultGraphConverter:
         for docs see https://graphneural.network/creating-dataset/
         """
         return CustomSpektralDataset(graphs=self.to_spektral_graphs())
+
+    def _verify_feature_funcs(self, funcs, feature_type: Literal["edge", "node"]):
+        for i, func in enumerate(funcs):
+            # Check if it has the attributes added by the decorator
+            if not hasattr(func, "feature_type"):
+                func_str = inspect.getsource(func).strip()
+                raise Exception(
+                    f"Error processing feature function:\n"
+                    f"{func.__name__} defined as:\n"
+                    f"{func_str}\n\n"
+                    "Function is missing the @graph_feature decorator. "
+                )
+
+            if func.feature_type != feature_type:
+                func_str = inspect.getsource(func).strip()
+                raise Exception(
+                    f"Error processing feature function:\n"
+                    f"{func.__name__} defined as:\n"
+                    f"{func_str}\n\n"
+                    "Function has an incorrect feature type edge features should be 'edge', node features should be 'node'. "
+                )
+
+    @property
+    def return_dtypes(self):
+        return pl.Struct(
+            {
+                "e": pl.List(pl.List(pl.Float64)),
+                "x": pl.List(pl.List(pl.Float64)),
+                "a": pl.List(pl.List(pl.Float64)),
+                "e_shape_0": pl.Int64,
+                "e_shape_1": pl.Int64,
+                "x_shape_0": pl.Int64,
+                "x_shape_1": pl.Int64,
+                "a_shape_0": pl.Int64,
+                "a_shape_1": pl.Int64,
+                self.graph_id_column: pl.String,
+                self.label_column: pl.Int64,
+            }
+        )
+
+    def to_graph_frames(self) -> List[dict]:
+        def process_chunk(chunk: pl.DataFrame) -> List[dict]:
+            return [
+                {
+                    "a": make_sparse(
+                        reshape_from_size(
+                            chunk["a"][i], chunk["a_shape_0"][i], chunk["a_shape_1"][i]
+                        )
+                    ),
+                    "x": reshape_from_size(
+                        chunk["x"][i], chunk["x_shape_0"][i], chunk["x_shape_1"][i]
+                    ),
+                    "e": reshape_from_size(
+                        chunk["e"][i], chunk["e_shape_0"][i], chunk["e_shape_1"][i]
+                    ),
+                    "y": np.asarray([chunk[self.label_column][i]]),
+                    "id": chunk[self.graph_id_column][i],
+                }
+                for i in range(len(chunk))
+            ]
+
+        graph_df = self._convert()
+        self.graph_frames = [
+            graph
+            for chunk in graph_df.lazy()
+            .collect(engine="gpu")
+            .iter_slices(self.chunk_size)
+            for graph in process_chunk(chunk)
+        ]
+        return self.graph_frames
